layout: move to style sub-folder
