+++ crates/termoxide_layout/src/style/box_model/padding.rs
@@ -10,11 +10,11 @@
 //!
 //! - It **forbids negative values** (unlike [`super::margin::Margin`]) — there
 //!   is no semantic meaning to "negative inner spacing" in CSS.
-//! - It has **no `auto` keyword** (unlike [`super::margin::Margin`]) — the
-//!   only intrinsic value for padding is zero.
+//! - It has **no `auto` keyword** (unlike [`super::margin::Margin`]) — the only
+//!   intrinsic value for padding is zero.
 
 use super::edges::Edges;
-use crate::unit::Unit;
+use crate::style::unit::Unit;
 
 /// Inner spacing between the border and the content. CSS `padding`.
 ///
@@ -31,12 +31,16 @@ use crate::unit::Unit;
 /// # Examples
 ///
 /// ```rust
-/// use termoxide_layout::box_model::Padding;
-/// use termoxide_layout::unit::Unit;
+/// use termoxide_layout::style::{box_model::Padding, unit::Unit};
 ///
 /// let p = Padding::all(Unit::cells(1));
 /// let p = Padding::symmetric(Unit::cells(1), Unit::cells(2));
-/// let p = Padding::new(Unit::cells(1), Unit::cells(2), Unit::cells(1), Unit::cells(2));
+/// let p = Padding::new(
+///     Unit::cells(1),
+///     Unit::cells(2),
+///     Unit::cells(1),
+///     Unit::cells(2),
+/// );
 /// assert_eq!(p.edges().left, Unit::cells(2));
 ///
 /// // Negative cells are clamped:
@@ -56,7 +60,7 @@ const fn css_padding_value(v: Unit) -> Unit {
             } else {
                 Unit::Cells(n)
             }
-        }
+        },
         Unit::Percent(_) => v,
         _ => Unit::ZERO,
     }
@@ -65,9 +69,7 @@ const fn css_padding_value(v: Unit) -> Unit {
 impl Padding {
     pub const ZERO: Self = Self(Edges::all(Unit::ZERO));
 
-    pub const fn all(v: Unit) -> Self {
-        Self(Edges::all(css_padding_value(v)))
-    }
+    pub const fn all(v: Unit) -> Self { Self(Edges::all(css_padding_value(v))) }
 
     pub const fn symmetric(vertical: Unit, horizontal: Unit) -> Self {
         Self(Edges::symmetric(
@@ -86,19 +88,13 @@ impl Padding {
     }
 
     /// Borrow the underlying [`Edges`] for read access.
-    pub const fn edges(&self) -> &Edges {
-        &self.0
-    }
+    pub const fn edges(&self) -> &Edges { &self.0 }
 
     /// Consume the wrapper and return the underlying [`Edges`].
-    pub const fn into_edges(self) -> Edges {
-        self.0
-    }
+    pub const fn into_edges(self) -> Edges { self.0 }
 }
 
 impl From<Edges> for Padding {
     /// Lift a raw [`Edges`] into a `Padding`, normalising each side.
-    fn from(e: Edges) -> Self {
-        Self::new(e.top, e.right, e.bottom, e.left)
-    }
+    fn from(e: Edges) -> Self { Self::new(e.top, e.right, e.bottom, e.left) }
 }
