+++ crates/termoxide_layout/src/style/color.rs
@@ -4,11 +4,11 @@ use std::hash::Hash;
 ///
 /// Terminal color support comes in three tiers:
 ///
-/// | Tier   | Type              | Support                                      |
-/// |--------|-------------------|----------------------------------------------|
-/// | 3/4-bit | [`NamedColor`]   | Universal — every terminal                   |
-/// | 8-bit  | [`Color::Indexed`] | xterm-256color and above                    |
-/// | 24-bit | [`Color::Rgb`]    | Modern terminals (kitty, iTerm2, WinTerm…)   |
+/// | Tier   | Type              | Support                                    |
+/// |--------|-------------------|--------------------------------------------|
+/// | 3/4-bit | [`NamedColor`]   | Universal — every terminal                 |
+/// | 8-bit  | [`Color::Indexed`] | xterm-256color and above                  |
+/// | 24-bit | [`Color::Rgb`]    | Modern terminals (kitty, iTerm2, WinTerm…) |
 ///
 /// When converting to `ratatui::style::Color`, degrade gracefully:
 /// prefer `Rgb`, fall back to `Indexed`, fall back to `Named`.
@@ -16,11 +16,11 @@ use std::hash::Hash;
 /// # Examples
 ///
 /// ```rust
-/// use termoxide_layout::color::{Color, NamedColor};
-/// let red   = Color::Named(NamedColor::Red);
+/// use termoxide_layout::style::color::{Color, NamedColor};
+/// let red = Color::Named(NamedColor::Red);
 /// # #[cfg(feature = "future")] {
 /// let coral = Color::rgb(255, 127, 80);
-/// let grey  = Color::indexed(240);
+/// let grey = Color::indexed(240);
 /// # }
 /// ```
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
@@ -53,14 +53,10 @@ pub enum Color {
 
 impl Color {
     #[cfg(feature = "future")]
-    pub const fn rgb(r: u8, g: u8, b: u8) -> Self {
-        Self::Rgb(r, g, b)
-    }
+    pub const fn rgb(r: u8, g: u8, b: u8) -> Self { Self::Rgb(r, g, b) }
 
     #[cfg(feature = "future")]
-    pub const fn indexed(i: u8) -> Self {
-        Self::Indexed(i)
-    }
+    pub const fn indexed(i: u8) -> Self { Self::Indexed(i) }
 
     /// Parse a `#RRGGBB` hex color at compile time.
     ///
@@ -77,7 +73,7 @@ impl Color {
                     (Some(r), Some(g), Some(b)) => Some(Self::Rgb(r, g, b)),
                     _ => None,
                 }
-            }
+            },
             _ => None,
         }
     }
@@ -152,9 +148,7 @@ pub enum NamedColor {
 
 impl NamedColor {
     /// The ANSI palette index (0–15) for this color.
-    pub const fn ansi_index(self) -> u8 {
-        self as u8
-    }
+    pub const fn ansi_index(self) -> u8 { self as u8 }
 
     #[cfg(feature = "ratatui")]
     pub fn to_ratatui(self) -> ratatui::style::Color {
