+++ crates/termoxide_layout/src/style/str.rs
@@ -1,5 +1,4 @@
-use std::borrow::Cow;
-use std::hash::Hash;
+use std::{borrow::Cow, hash::Hash};
 
 /// A CSS-like string value.
 ///
@@ -10,7 +9,7 @@ use std::hash::Hash;
 ///
 /// The inner `Cow<'static, str>` means proc_macro-generated code like:
 /// ```rust
-/// use termoxide_layout::str::Str;
+/// use termoxide_layout::style::str::Str;
 /// let font = Str::from_static("JetBrains Mono");
 /// ```
 /// involves **zero heap allocation** — the slice lives in the binary's
@@ -23,8 +22,8 @@ use std::hash::Hash;
 /// # Examples
 ///
 /// ```rust
-/// use termoxide_layout::str::Str;
-/// let a: Str = "monospace".into();              // static borrow, no alloc
+/// use termoxide_layout::style::str::Str;
+/// let a: Str = "monospace".into(); // static borrow, no alloc
 /// let b = Str::from_string(format!("Font-{}", 42)); // heap-allocated
 /// ```
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Default)]
@@ -34,40 +33,26 @@ impl Str {
     /// Construct from a `'static` str — zero allocation.
     ///
     /// Preferred for proc_macro output.
-    pub const fn from_static(s: &'static str) -> Self {
-        Self(Cow::Borrowed(s))
-    }
+    pub const fn from_static(s: &'static str) -> Self { Self(Cow::Borrowed(s)) }
 
     /// Construct from a runtime-owned `String` — heap-allocates.
-    pub fn from_string(s: String) -> Self {
-        Self(Cow::Owned(s))
-    }
+    pub fn from_string(s: String) -> Self { Self(Cow::Owned(s)) }
 
     /// Borrow the inner string slice.
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
+    pub fn as_str(&self) -> &str { &self.0 }
 
     /// Returns `true` if the string is empty.
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
+    pub fn is_empty(&self) -> bool { self.0.is_empty() }
 }
 
 impl From<&'static str> for Str {
-    fn from(s: &'static str) -> Self {
-        Self::from_static(s)
-    }
+    fn from(s: &'static str) -> Self { Self::from_static(s) }
 }
 impl From<String> for Str {
-    fn from(s: String) -> Self {
-        Self::from_string(s)
-    }
+    fn from(s: String) -> Self { Self::from_string(s) }
 }
 impl AsRef<str> for Str {
-    fn as_ref(&self) -> &str {
-        self.as_str()
-    }
+    fn as_ref(&self) -> &str { self.as_str() }
 }
 
 impl std::fmt::Display for Str {
